Eve/requirements-application.txt
openai==0.27.4
haystack==0.42

tiktoken==0.3.3

weaviate-client==3.15.6

aiohttp==3.8.4
aiodns==3.0.0
apify-client==1.0.0

python-dotenv==1.0.0
